pyproject.toml
[tool.poetry]
name = "python-accounting"
version = "1.0.0"
description = "The world's most comprehensive python Double Entry Accounting Library"
authors = ["Edward Mungai <emungai@microbooks.io>"]
readme = "README.md"
packages = [{include = "python_accounting"}, {include = "config.toml"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.23"
python-dateutil = "^2.8.2"
strenum = "^0.4.15"
toml = "^0.10.2"
mysqlclient = "^2.2.4"
psycopg2 = "^2.9.9"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"

[tool.poetry.group.dev.dependencies]
sphinx-autoapi = "^3.0.0"
sphinx-rtd-theme = "^2.0.0"
sphinx = "^7.2.6"
myst-parser = "^2.0.0"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
